Removed "order" state and restructured class.
Removed the initialisation of the "order" state from the init method. Class methods can no longer iterate up to "self.order" or "polynomial.order". Instead, the "for" loops in class methods now iterate over the proper values up to the length of the "coefficients" state. Fixes coefficient dropping when using the addPolynomial(), derivative(), and integral() methods.